\name{checkETPError}
\alias{checkETPError}
\title{Checks for the presence of an error in the returned JSON and prints a warning.}
\usage{checkETPError(input)}
\description{Checks for the presence of an error in the returned JSON and prints a warning.}
\value{none -- warns if error found}
\author{Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}}
\arguments{\item{input}{the list which should be checked}}
