\name{etp.getGene}
\alias{etp.getGene}
\title{Gets information about a particular gene...}
\usage{etp.getGene(conn, entrezID)}
\description{Gets information about a particular gene}
\value{information about the gene including its name, description, symbols, and accession numbers.}
\author{Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}}
\arguments{\item{conn}{the connection object generated by etp.connect()}
\item{entrezID}{the EntrezID of the gene in which you're interested}}
