\name{etp.getPlatforms}
\alias{etp.getPlatforms}
\title{Gets a list of the available platforms in the database...}
\usage{etp.getPlatforms(conn)}
\description{Gets a list of the available platforms in the database}
\value{a list of the available platforms with their IDs as the row.names}
\author{Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}}
\arguments{\item{conn}{the connection object generated by etp.connect()}}
