\name{pm.getPlatforms}
\alias{pm.getPlatforms}
\title{Gets a list of the available platforms in the database...}
\usage{pm.getPlatforms(conn)}
\description{Gets a list of the available platforms in the database}
\value{a list of the available platforms with their IDs as the row.names}
\author{Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}}
\arguments{\item{conn}{the connection object generated by pm.connect()}}
