\name{pm.getProbe}
\alias{pm.getProbe}
\title{Gets information about a particular probe...}
\usage{pm.getProbe(conn, probeID, platformID, probeName)}
\description{Gets information about a particular probe}
\value{information about the probe specified including its probe ID, platformID, and probe name.}
\author{Jeffrey D. Allen \email{Jeffrey.Allen@UTSouthwestern.edu}}
\arguments{\item{conn}{the connection object generated by pm.connect()}
\item{probeID}{the internal probeID used to uniquely reference probes}
\item{platformID}{the ID of the platform on which this probeName is found.}
\item{probeName}{the vendor-supplied name of the probe for this platformID.}}
